save comment: saves a cooment[undone]
