+++ server/app.js
@@ -1,12 +1,12 @@
 import express from 'express';
-import createblogRouters from "./routers/createblogRouters";
+import commentRouters from "./routers/commentRouters";
 require ('../server/config/mongoose')
 const app = express();
 
 app.use(express.json())
 app.use(express.urlencoded({extended:false}))
 
-app.use(createblogRouters)
+app.use(commentRouters)
