--- server/routers/createblogRouters.js
@@ -1,11 +1,0 @@
-import express from 'express';
-import createblogControllers from "../controllers/createblogControllers";
-
-const createblogRouters = express.Router();
-
-
-
-createblogRouters.post('/blogs',createblogControllers.save)
-
-
-export default createblogRouters
